Version 3 of spindle_mount2, trying to pre-balance it
This version, like v2, is meant to be attached to a strongly built axle/hub,
trying for a rigid build rather than light-weight.

The idea to pre-balance is to compute the center-of-mass and inertia tensor
of the slic3r-extruded model, and then add appropriate cutouts/extras to get
as good a balance in the model as printed, reducing the need for extra
balance weights after printing.